+++ src/ua/data_types/status_code.rs
@@ -1,4 +1,7 @@
-use open62541_sys::{UA_StatusCode, UA_STATUSCODE_GOOD};
+use open62541_sys::{
+    UA_StatusCode, UA_StatusCode_isBad, UA_StatusCode_isGood, UA_StatusCode_isUncertain,
+    UA_STATUSCODE_GOOD,
+};
 
 crate::data_type!(StatusCode);
 
@@ -13,11 +16,41 @@ impl StatusCode {
     }
 
     /// Checks if status code is good.
+    ///
+    /// Good status codes indicate that the operation was successful and the associated results may
+    /// be used.
+    ///
+    /// Note: This only checks the _severity_ of the status code. If you want to see if the code is
+    /// exactly the single status code [`GOOD`](Self::GOOD), use comparison instead:
+    ///
+    /// ```rust
+    /// use open62541::ua;
+    ///
+    /// # let status_code = ua::StatusCode::GOOD;
+    /// if status_code == ua::StatusCode::GOOD {
+    ///     //
+    /// }
+    /// ````
     #[must_use]
-    pub const fn is_good(&self) -> bool {
-        // TODO: Check name of this method. Consider potential clash with `UA_StatusCode_isGood()`
-        // which only makes check for _severity_ of status code (i.e. may match an entire range of
-        // codes).
-        self.0 == UA_STATUSCODE_GOOD
+    pub fn is_good(&self) -> bool {
+        unsafe { UA_StatusCode_isGood(self.0) }
+    }
+
+    /// Checks if status code is uncertain.
+    ///
+    /// Uncertain status codes indicate that the operation was partially successful and that
+    /// associated results might not be suitable for some purposes.
+    #[must_use]
+    pub fn is_uncertain(&self) -> bool {
+        unsafe { UA_StatusCode_isUncertain(self.0) }
+    }
+
+    /// Checks if status code is bad.
+    ///
+    /// Bad status codes indicate that the operation failed and any associated results cannot be
+    /// used.
+    #[must_use]
+    pub fn is_bad(&self) -> bool {
+        unsafe { UA_StatusCode_isBad(self.0) }
     }
 }